+++ bigchaindb/backend/utils.py
